# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Júlia Vendedora - AI sales agent for CaptaClin
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type LeadFunnelStage = typeof LeadFunnelStage[keyof typeof LeadFunnelStage];


export const LeadFunnelStage = {
  new: 'new',
  contacted: 'contacted',
  qualified: 'qualified',
  interested: 'interested',
  objection: 'objection',
  closing: 'closing',
  closed: 'closed',
  lost: 'lost',
} as const;

/**
 * @nullable
 */
export type LeadPlanInterest = typeof LeadPlanInterest[keyof typeof LeadPlanInterest] | null;


export const LeadPlanInterest = {
  basic: 'basic',
  essencial: 'essencial',
  pro: 'pro',
} as const;

export type LeadStatus = typeof LeadStatus[keyof typeof LeadStatus];


export const LeadStatus = {
  hot: 'hot',
  warm: 'warm',
  cold: 'cold',
  closed: 'closed',
  lost: 'lost',
} as const;

export interface Lead {
  id: number;
  /** @nullable */
  name?: string | null;
  phone: string;
  /** @nullable */
  origin?: string | null;
  funnelStage: LeadFunnelStage;
  /** @nullable */
  painPoints?: string | null;
  /** @nullable */
  mainObjection?: string | null;
  /** @nullable */
  planInterest?: LeadPlanInterest;
  status: LeadStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  /** @nullable */
  pausedUntil?: string | null;
  handoffRequested?: boolean;
  createdAt: string;
  updatedAt: string;
}

export type LeadUpdateFunnelStage = typeof LeadUpdateFunnelStage[keyof typeof LeadUpdateFunnelStage];


export const LeadUpdateFunnelStage = {
  new: 'new',
  contacted: 'contacted',
  qualified: 'qualified',
  interested: 'interested',
  objection: 'objection',
  closing: 'closing',
  closed: 'closed',
  lost: 'lost',
} as const;

export type LeadUpdatePlanInterest = typeof LeadUpdatePlanInterest[keyof typeof LeadUpdatePlanInterest];


export const LeadUpdatePlanInterest = {
  basic: 'basic',
  essencial: 'essencial',
  pro: 'pro',
} as const;

export type LeadUpdateStatus = typeof LeadUpdateStatus[keyof typeof LeadUpdateStatus];


export const LeadUpdateStatus = {
  hot: 'hot',
  warm: 'warm',
  cold: 'cold',
  closed: 'closed',
  lost: 'lost',
} as const;

export interface LeadUpdate {
  name?: string;
  funnelStage?: LeadUpdateFunnelStage;
  painPoints?: string;
  mainObjection?: string;
  planInterest?: LeadUpdatePlanInterest;
  status?: LeadUpdateStatus;
  notes?: string;
  handoffRequested?: boolean;
  /** @nullable */
  pausedUntil?: null;
}

export interface LeadListResponse {
  leads: Lead[];
  total: number;
}

export type MessageDirection = typeof MessageDirection[keyof typeof MessageDirection];


export const MessageDirection = {
  inbound: 'inbound',
  outbound: 'outbound',
} as const;

export type MessageMessageType = typeof MessageMessageType[keyof typeof MessageMessageType];


export const MessageMessageType = {
  text: 'text',
  audio: 'audio',
  image: 'image',
} as const;

export interface Message {
  id: number;
  leadId: number;
  direction: MessageDirection;
  content: string;
  messageType: MessageMessageType;
  createdAt: string;
}

export type FollowUpStatus = typeof FollowUpStatus[keyof typeof FollowUpStatus];


export const FollowUpStatus = {
  pending: 'pending',
  sent: 'sent',
  cancelled: 'cancelled',
} as const;

export interface FollowUp {
  id: number;
  leadId: number;
  scheduledAt: string;
  touchNumber: number;
  /** @nullable */
  messageTemplate?: string | null;
  status: FollowUpStatus;
  createdAt: string;
}

export interface DashboardStats {
  totalLeads: number;
  hotLeads: number;
  closedLeads: number;
  lostLeads: number;
  conversionRate: number;
  pendingFollowups: number;
  handoffsPending: number;
}

export interface FunnelStage {
  stage: string;
  count: number;
}

export interface ActivityItem {
  leadId: number;
  /** @nullable */
  leadName: string | null;
  phone: string;
  event: string;
  funnelStage: string;
  status: string;
  timestamp: string;
}

export type WebhookPayloadData = { [key: string]: unknown };

export interface WebhookPayload {
  event?: string;
  instance?: string;
  data?: WebhookPayloadData;
}

export interface WebhookAck {
  ok: boolean;
}

export type ListLeadsParams = {
status?: ListLeadsStatus;
funnelStage?: ListLeadsFunnelStage;
search?: string;
limit?: number;
offset?: number;
};

export type ListLeadsStatus = typeof ListLeadsStatus[keyof typeof ListLeadsStatus];


export const ListLeadsStatus = {
  hot: 'hot',
  warm: 'warm',
  cold: 'cold',
  closed: 'closed',
  lost: 'lost',
} as const;

export type ListLeadsFunnelStage = typeof ListLeadsFunnelStage[keyof typeof ListLeadsFunnelStage];


export const ListLeadsFunnelStage = {
  new: 'new',
  contacted: 'contacted',
  qualified: 'qualified',
  interested: 'interested',
  objection: 'objection',
  closing: 'closing',
  closed: 'closed',
  lost: 'lost',
} as const;

export type GetRecentActivityParams = {
limit?: number;
};