# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Júlia Vendedora - AI sales agent for CaptaClin
 * OpenAPI spec version: 0.1.0
 */
import type { LeadFunnelStage } from './leadFunnelStage';
import type { LeadPlanInterest } from './leadPlanInterest';
import type { LeadStatus } from './leadStatus';

export interface Lead {
  id: number;
  /** @nullable */
  name?: string | null;
  phone: string;
  /** @nullable */
  origin?: string | null;
  funnelStage: LeadFunnelStage;
  /** @nullable */
  painPoints?: string | null;
  /** @nullable */
  mainObjection?: string | null;
  /** @nullable */
  planInterest?: LeadPlanInterest;
  status: LeadStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  /** @nullable */
  pausedUntil?: string | null;
  handoffRequested?: boolean;
  createdAt: string;
  updatedAt: string;
}