# lib/api-zod/src/generated/types/leadUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Júlia Vendedora - AI sales agent for CaptaClin
 * OpenAPI spec version: 0.1.0
 */
import type { LeadUpdateFunnelStage } from './leadUpdateFunnelStage';
import type { LeadUpdatePlanInterest } from './leadUpdatePlanInterest';
import type { LeadUpdateStatus } from './leadUpdateStatus';

export interface LeadUpdate {
  name?: string;
  funnelStage?: LeadUpdateFunnelStage;
  painPoints?: string;
  mainObjection?: string;
  planInterest?: LeadUpdatePlanInterest;
  status?: LeadUpdateStatus;
  notes?: string;
  handoffRequested?: boolean;
  /** @nullable */
  pausedUntil?: null;
}